Разделяет утилитный модуль
